pyproject.toml
[project]
name = "this-day"
version = "0.1.0"
description = ""
authors = [
    {name = "t0nylombardi",email = "alombardi.331@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pendulum (>=3.1.0,<4.0.0)",
    "rich-pyfiglet (>=1.0.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "emoji (>=2.15.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "this"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
